+++ test/unit/events.test.js
@@ -21,23 +21,12 @@
 
 const { test } = require('tap')
 const semver = require('semver')
-const intoStream = require('into-stream')
-const { Client, Connection, events } = require('../../index')
+const { Client, events } = require('../../index')
+const { TimeoutError } = require('../../lib/errors')
 const {
-  TimeoutError,
-  ConnectionError,
-  ResponseError,
-  RequestAbortedError,
-  SerializationError,
-  DeserializationError
-} = require('../../lib/errors')
-const {
-  buildServer,
   connection: {
     MockConnection,
-    MockConnectionError,
-    MockConnectionTimeout,
-    buildMockConnection
+    MockConnectionTimeout
   }
 } = require('../utils')
 
@@ -305,429 +294,4 @@ test('Emit event', t => {
   })
 
   client.emit(events.REQUEST, null, { hello: 'world' })
-})
-
-test('Event order', t => {
-  t.test('No errors', t => {
-    t.plan(10)
-
-    const client = new Client({
-      node: 'http://localhost:9200',
-      Connection: MockConnection
-    })
-
-    const order = [
-      events.SERIALIZATION,
-      events.REQUEST,
-      events.DESERIALIZATION,
-      events.RESPONSE
-    ]
-
-    client.on(events.SERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.SERIALIZATION)
-    })
-
-    client.on(events.REQUEST, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.REQUEST)
-    })
-
-    client.on(events.DESERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.DESERIALIZATION)
-    })
-
-    client.on(events.RESPONSE, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.RESPONSE)
-    })
-
-    client.info((err, result) => {
-      t.error(err)
-      t.strictEqual(order.length, 0)
-    })
-  })
-
-  t.test('Connection error', t => {
-    t.plan(10)
-
-    const client = new Client({
-      node: 'http://localhost:9200',
-      Connection: MockConnectionError,
-      maxRetries: 1
-    })
-
-    const order = [
-      events.SERIALIZATION,
-      events.REQUEST,
-      events.REQUEST,
-      events.RESPONSE
-    ]
-
-    client.on(events.SERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.SERIALIZATION)
-    })
-
-    client.on(events.REQUEST, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.REQUEST)
-    })
-
-    client.on(events.DESERIALIZATION, (_err, request) => {
-      t.fail('Should not be called')
-    })
-
-    client.on(events.RESPONSE, (err, request) => {
-      t.ok(err instanceof ConnectionError)
-      t.strictEqual(order.shift(), events.RESPONSE)
-    })
-
-    client.info((err, result) => {
-      t.ok(err instanceof ConnectionError)
-      t.strictEqual(order.length, 0)
-    })
-  })
-
-  t.test('TimeoutError error', t => {
-    t.plan(10)
-
-    const client = new Client({
-      node: 'http://localhost:9200',
-      Connection: MockConnectionTimeout,
-      maxRetries: 1
-    })
-
-    const order = [
-      events.SERIALIZATION,
-      events.REQUEST,
-      events.REQUEST,
-      events.RESPONSE
-    ]
-
-    client.on(events.SERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.SERIALIZATION)
-    })
-
-    client.on(events.REQUEST, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.REQUEST)
-    })
-
-    client.on(events.DESERIALIZATION, (_err, request) => {
-      t.fail('Should not be called')
-    })
-
-    client.on(events.RESPONSE, (err, request) => {
-      t.ok(err instanceof TimeoutError)
-      t.strictEqual(order.shift(), events.RESPONSE)
-    })
-
-    client.info((err, result) => {
-      t.ok(err instanceof TimeoutError)
-      t.strictEqual(order.length, 0)
-    })
-  })
-
-  t.test('RequestAbortedError error', t => {
-    t.plan(8)
-
-    const client = new Client({
-      node: 'http://localhost:9200',
-      Connection: MockConnectionTimeout,
-      maxRetries: 1
-    })
-
-    const order = [
-      events.SERIALIZATION,
-      events.REQUEST,
-      events.RESPONSE
-    ]
-
-    client.on(events.SERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.SERIALIZATION)
-    })
-
-    client.on(events.REQUEST, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.REQUEST)
-    })
-
-    client.on(events.DESERIALIZATION, (_err, request) => {
-      t.fail('Should not be called')
-    })
-
-    client.on(events.RESPONSE, (err, request) => {
-      t.ok(err instanceof RequestAbortedError)
-      t.strictEqual(order.shift(), events.RESPONSE)
-    })
-
-    const request = client.info((err, result) => {
-      t.ok(err instanceof RequestAbortedError)
-      t.strictEqual(order.length, 0)
-    })
-
-    request.abort()
-  })
-
-  t.test('ResponseError error (no retry)', t => {
-    t.plan(10)
-
-    const MockConnection = buildMockConnection({
-      onRequest (params) {
-        return {
-          statusCode: 400,
-          body: { hello: 'world' }
-        }
-      }
-    })
-
-    const client = new Client({
-      node: 'http://localhost:9200',
-      Connection: MockConnection,
-      maxRetries: 1
-    })
-
-    const order = [
-      events.SERIALIZATION,
-      events.REQUEST,
-      events.DESERIALIZATION,
-      events.RESPONSE
-    ]
-
-    client.on(events.SERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.SERIALIZATION)
-    })
-
-    client.on(events.REQUEST, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.REQUEST)
-    })
-
-    client.on(events.DESERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.DESERIALIZATION)
-    })
-
-    client.on(events.RESPONSE, (err, request) => {
-      t.ok(err instanceof ResponseError)
-      t.strictEqual(order.shift(), events.RESPONSE)
-    })
-
-    client.info((err, result) => {
-      t.ok(err instanceof ResponseError)
-      t.strictEqual(order.length, 0)
-    })
-  })
-
-  t.test('ResponseError error (with retry)', t => {
-    t.plan(14)
-
-    const MockConnection = buildMockConnection({
-      onRequest (params) {
-        return {
-          statusCode: 504,
-          body: { hello: 'world' }
-        }
-      }
-    })
-
-    const client = new Client({
-      node: 'http://localhost:9200',
-      Connection: MockConnection,
-      maxRetries: 1
-    })
-
-    const order = [
-      events.SERIALIZATION,
-      events.REQUEST,
-      events.DESERIALIZATION,
-      events.REQUEST,
-      events.DESERIALIZATION,
-      events.RESPONSE
-    ]
-
-    client.on(events.SERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.SERIALIZATION)
-    })
-
-    client.on(events.REQUEST, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.REQUEST)
-    })
-
-    client.on(events.DESERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.DESERIALIZATION)
-    })
-
-    client.on(events.RESPONSE, (err, request) => {
-      t.ok(err instanceof ResponseError)
-      t.strictEqual(order.shift(), events.RESPONSE)
-    })
-
-    client.info((err, result) => {
-      t.ok(err instanceof ResponseError)
-      t.strictEqual(order.length, 0)
-    })
-  })
-
-  t.test('Serialization Error', t => {
-    t.plan(6)
-
-    const client = new Client({
-      node: 'http://localhost:9200',
-      Connection: MockConnection,
-      maxRetries: 1
-    })
-
-    const order = [
-      events.SERIALIZATION,
-      events.REQUEST
-    ]
-
-    client.on(events.SERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.SERIALIZATION)
-    })
-
-    client.on(events.REQUEST, (err, request) => {
-      t.ok(err instanceof SerializationError)
-      t.strictEqual(order.shift(), events.REQUEST)
-    })
-
-    client.on(events.DESERIALIZATION, (_err, request) => {
-      t.fail('Should not be called')
-    })
-
-    client.on(events.RESPONSE, (_err, request) => {
-      t.fail('Should not be called')
-    })
-
-    const body = {}
-    body.o = body
-    client.index({ index: 'test', body }, (err, result) => {
-      t.ok(err instanceof SerializationError)
-      t.strictEqual(order.length, 0)
-    })
-  })
-
-  t.test('Deserialization Error', t => {
-    t.plan(10)
-
-    class MockConnection extends Connection {
-      request (params, callback) {
-        const body = '{"hello":"wor'
-        const stream = intoStream(body)
-        stream.statusCode = 200
-        stream.headers = {
-          'content-type': 'application/json;utf=8',
-          'content-length': body.length,
-          connection: 'keep-alive',
-          date: new Date().toISOString()
-        }
-        stream.on('close', () => t.pass('Stream destroyed'))
-        process.nextTick(callback, null, stream)
-        return { abort () {} }
-      }
-    }
-
-    const client = new Client({
-      node: 'http://localhost:9200',
-      Connection: MockConnection,
-      maxRetries: 1
-    })
-
-    const order = [
-      events.SERIALIZATION,
-      events.REQUEST,
-      events.DESERIALIZATION,
-      events.RESPONSE
-    ]
-
-    client.on(events.SERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.SERIALIZATION)
-    })
-
-    client.on(events.REQUEST, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.REQUEST)
-    })
-
-    client.on(events.DESERIALIZATION, (err, request) => {
-      t.error(err)
-      t.strictEqual(order.shift(), events.DESERIALIZATION)
-    })
-
-    client.on(events.RESPONSE, (err, request) => {
-      t.ok(err instanceof DeserializationError)
-      t.strictEqual(order.shift(), events.RESPONSE)
-    })
-
-    client.info((err, result) => {
-      t.ok(err instanceof DeserializationError)
-      t.strictEqual(order.length, 0)
-    })
-  })
-
-  t.test('Socket destroyed while reading the body', t => {
-    t.plan(14)
-
-    function handler (req, res) {
-      const body = JSON.stringify({ hello: 'world' })
-      res.setHeader('Content-Type', 'application/json;utf=8')
-      res.setHeader('Content-Length', body.length + '')
-      res.write(body.slice(0, -5))
-      setTimeout(() => {
-        res.socket.destroy()
-      }, 500)
-    }
-
-    buildServer(handler, ({ port }, server) => {
-      const client = new Client({ node: `http://localhost:${port}`, maxRetries: 1 })
-
-      const order = [
-        events.SERIALIZATION,
-        events.REQUEST,
-        events.DESERIALIZATION,
-        events.REQUEST,
-        events.DESERIALIZATION,
-        events.RESPONSE
-      ]
-
-      client.on(events.SERIALIZATION, (err, request) => {
-        t.error(err)
-        t.strictEqual(order.shift(), events.SERIALIZATION)
-      })
-
-      client.on(events.REQUEST, (err, request) => {
-        t.error(err)
-        t.strictEqual(order.shift(), events.REQUEST)
-      })
-
-      client.on(events.DESERIALIZATION, (err, request) => {
-        t.error(err)
-        t.strictEqual(order.shift(), events.DESERIALIZATION)
-      })
-
-      client.on(events.RESPONSE, (err, request) => {
-        t.ok(err instanceof ConnectionError)
-        t.strictEqual(order.shift(), events.RESPONSE)
-      })
-
-      client.info((err, result) => {
-        t.ok(err instanceof ConnectionError)
-        t.strictEqual(order.length, 0)
-        server.stop()
-      })
-    })
-  })
-
-  t.end()
 })
